+++ src/minecraft/co/uk/flansmods/common/CommonProxy.java
@@ -231,7 +231,7 @@ public class CommonProxy
 		}
 		
 		//Find the stack of engines that is fastest but which also has enough for this driveable
-		int bestEngineSpeed = -1;
+		float bestEngineSpeed = -1F;
 		ItemStack bestEngineStack = null;
 		for(PartType part : engines.keySet())
 		{
